# Lab2/pzpi-22-2-moskovchenko-yevhenii-task2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StaffFlow api",
    "version": "1.0"
  },
  "paths": {
    "/api/ChangeRequests": {
      "get": {
        "tags": [
          "ChangeRequests"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ChangeRequests"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ChangeRequests/{id}": {
      "get": {
        "tags": [
          "ChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DayTypes": {
      "get": {
        "tags": [
          "DayTypes"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "DayTypes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DayTypes/{id}": {
      "get": {
        "tags": [
          "DayTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "DayTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DayType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "DayTypes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserChangeRequests": {
      "get": {
        "tags": [
          "UserChangeRequests"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "UserChangeRequests"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserChangeRequests/{id}": {
      "get": {
        "tags": [
          "UserChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "UserChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UserChangeRequests"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserWorkingDays": {
      "get": {
        "tags": [
          "UserWorkingDays"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "UserWorkingDays"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/UserWorkingDays/{id}": {
      "get": {
        "tags": [
          "UserWorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "UserWorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserWorkingDay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UserWorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WorkingDays": {
      "get": {
        "tags": [
          "WorkingDays"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "WorkingDays"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WorkingDays/{id}": {
      "get": {
        "tags": [
          "WorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "WorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingDay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WorkingDays"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangeRequest": {
        "type": "object",
        "properties": {
          "changeRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "requestDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "dayTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "dayType": {
            "$ref": "#/components/schemas/DayType"
          },
          "userChangeRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChangeRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DayType": {
        "type": "object",
        "properties": {
          "dayTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "dayTypeName": {
            "type": "string",
            "nullable": true
          },
          "workingDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkingDay"
            },
            "nullable": true
          },
          "changeRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChangeRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "userWorkingDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserWorkingDay"
            },
            "nullable": true
          },
          "userChangeRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserChangeRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserChangeRequest": {
        "type": "object",
        "properties": {
          "userChangeRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "changeRequestId": {
            "type": "integer",
            "format": "int32"
          },
          "changeRequest": {
            "$ref": "#/components/schemas/ChangeRequest"
          }
        },
        "additionalProperties": false
      },
      "UserWorkingDay": {
        "type": "object",
        "properties": {
          "userWorkingDayId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "workingDayId": {
            "type": "integer",
            "format": "int32"
          },
          "workingDay": {
            "$ref": "#/components/schemas/WorkingDay"
          }
        },
        "additionalProperties": false
      },
      "WorkingDay": {
        "type": "object",
        "properties": {
          "workingDayId": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-span"
          },
          "endTime": {
            "type": "string",
            "format": "date-span"
          },
          "dayTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "dayType": {
            "$ref": "#/components/schemas/DayType"
          },
          "userWorkingDays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserWorkingDay"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}